% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.agg.cdf.R
\name{annual.agg.cdf}
\alias{annual.agg.cdf}
\alias{yearly.agg.cdf}
\title{Mirror of \code{\link{cdf}}: probability distribution fitting with L-Moments}
\usage{
annual.agg.cdf(
  x,
  para,
  f.name = "f",
  aggr.name = NA,
  dd.name = NA,
  is_depth = FALSE,
  dd_formatter = NA,
  ...
)

yearly.agg.cdf(...)
}
\arguments{
\item{x}{vector of variable values (pe.g. precipitation intensity).}

\item{para}{object returned by \code{\link{annual.agg.pel}}.}

\item{f.name}{column name for the probabilities/frequancies. Default is \code{"f"}.}

\item{aggr.name, dd.name}{column names. see \code{\link{annual.agg.samlmu}} or \code{\link{annual.agg.pel}}.}

\item{is_depth}{logical. Default is \code{FALSE}. It must be set \code{TRUE} in case \code{x} refers to precipitation depth.}

\item{dd_formatter}{string formatter for duration in the function value (currently not used).}

\item{...}{further arguments.}
}
\value{
a data frame containg values, probabilities/frequencies and duration values
}
\description{
Mirror of \code{\link{cdf}}: probability distribution fitting with L-Moments
}
\examples{

library(RMAWGEN)
data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
                       PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

y <- annual.agg(x,dd=1:20,time) 
lmom <- annual.agg.samlmu(y)
## TO CONTUE
y <- annual.agg(x,dd=1:5,time) 
lmom <- annual.agg.samlmu(y)
lmrd(lmom)
z <- annual.agg.pel(distrib="gpa",x=y,lmom=lmom)
ff <- annual.agg.qua(f=c(0.5,0.9),para=z)
out <- annual.agg.cdf(x=ff$aggr,para=z)

out[out$f \%in\% ff$f,names(ff)]
ff
}
