% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.agg..conf.interval.R
\name{annual.agg.conf.interval}
\alias{annual.agg.conf.interval}
\alias{yearly.agg.conf.interval}
\title{Confidence Interval}
\usage{
annual.agg.conf.interval(
  x = "sample",
  para,
  f.name = "f",
  aggr.name = NA,
  dd.name = NA,
  jj.name = "jj",
  dd_formatter = NA,
  R = 9999,
  boot.ci.type = c("perc", "norm", "basic", "stud"),
  conf = 0.95,
  probability_distribution_attrname = "probability_distrib",
  extreme.f = c(0.99, 0.995),
  ...
)

yearly.agg.conf.interval(...)
}
\arguments{
\item{x}{sample or \code{x} argument. See \code{\link{annual.agg.cdf}}}

\item{para}{object returned by \code{\link{annual.agg.pel}}.}

\item{f.name}{column name for the probabilities/frequencies. Default is \code{"f"}.}

\item{aggr.name, dd.name}{column names. see \code{\link{annual.agg.samlmu}} or \code{\link{annual.agg.pel}}.}

\item{jj.name}{column name for plotting position}

\item{dd_formatter}{see \code{\link{annual.agg.cdf}}}

\item{R}{see \code{R} in \code{\link{boot}}}

\item{boot.ci.type, conf}{see \code{type,conf} in \code{\link{boot.ci}}}

\item{probability_distribution_attrname}{see \code{\link{qua}}}

\item{extreme.f}{last point parameter (documention to be completed)}

\item{...}{further arguments.}
}
\description{
Bootstrap confidence interval for  Probability / Frequency / Return Period Extimation (for event of a specific duration). Function experimental.
}
\examples{

library(RMAWGEN)
library(ggplot2)
data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
                       PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570
y <- annual.agg(x,dd=1,time) ## duration: 1 day 
##
y$aggr[y$aggr<0] <- NA
y <- y[!is.na(y$aggr),]

##

lmom <- annual.agg.samlmu(y)


z <- annual.agg.pel(distrib="gev",x=y,lmom=lmom)
####
out <- annual.agg.conf.interval(para=z) ##,extreme.f = NULL) ##,boot.ci.type="perc")
out$rt <- 1/(1-out$f)
out$rt_min <- 1/(1-out$f_high)
out$rt_max <- 1/(1-out$f_low)
out$rt_jj <- 1/(1-out$jj)
## ggplot plot 
gg <- ggplot()+geom_point(aes(x=rt_jj,y=aggr_obs),data=out)
gg <- gg+geom_line(aes(x=rt,y=aggr_obs),data=out)+theme_bw()
gg <- gg+geom_ribbon(aes(x=rt,ymax=aggr_high,ymin=aggr_low),data=out,fill="grey",alpha=0.5)
gg <- gg+scale_x_log10()
gg <- gg+xlab("Return Period [yr]")+ylab("Precipitation [mm]")
##
}
\seealso{
\code{\link{annual.agg.idf.samlmu}},\code{\link{annual.agg.qua}}
}
