% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.agg.idf.samlmu.R
\name{annual.agg.idf.samlmu}
\alias{annual.agg.idf.samlmu}
\alias{yearly.agg.idf.samlmu}
\title{Sample L-Moments of annual maxima or minima values}
\usage{
annual.agg.idf.samlmu(
  x,
  lmom = NULL,
  aggr.name = "aggr",
  dd.name = "dd",
  nn_moms = c("l_1", "l_2"),
  ...
)

yearly.agg.idf.samlmu(x, ...)
}
\arguments{
\item{x}{object returned by \code{\link{annual.agg}} or \code{\link{yearly.agg}}}

\item{lmom}{object returned by \code{\link{annual.agg.samlmu}} or \code{NULL}.}

\item{aggr.name, dd.name}{optional column mane for \code{x}. See function usage.}

\item{...}{further arguments for \code{\link{samlmu}}}
}
\value{
a data frame containing the L-moment of rainfall intensity for each duration category. LOREM IPSUM
}
\description{
Sample L-Moments of annual maxima or minima values
}
\examples{

library(RMAWGEN)
data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

y <- annual.agg(x,dd=1:10,time)
out <- annual.agg.idf.samlmu(y)



}
\seealso{
\code{\link{samlmu}]} , \code{\link{annual.agg.samlmu}}
}
