% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.agg.qua.R
\name{annual.agg.qua}
\alias{annual.agg.qua}
\alias{yearly.agg.qua}
\title{Mirror of \code{\link{qua}}: probability distribution fitting with L-Moments}
\usage{
annual.agg.qua(
  f = (0:10)/10,
  para,
  f.name = "f",
  aggr.name = NA,
  dd.name = NA,
  use_ggplot = TRUE,
  xlab = "duration",
  ylab = "value",
  fill = c("#ca0020", "#0571b0", "#bababa", "#bababb", "#bababc"),
  nrun = 5,
  ...
)

yearly.agg.qua(...)
}
\arguments{
\item{f}{vector of probabilities/frequencies.}

\item{para}{object returned by \code{\link{annual.agg.pel}}.}

\item{f.name}{column name for the probabilities/frequancies. Deafault is \code{"f"}.}

\item{aggr.name, dd.name}{column names. see \code{\link{annual.agg.samlmu}} or \code{\link{annual.agg.pel}}.}

\item{use_ggplot}{logical. if iy is \code{TRUE} (dafault) boxplots are created. See function result descriptions.}

\item{xlab, ylab, fill}{arguments for \code{ggplot2} functions. It is used if \code{use_ggplot==TRUE}. See \code{\link{geom_boxplot}},\code{\link{scale_fill_manual}},\code{\link{labs}}}

\item{nrun}{additional Monte-Carlo generations. It is used if \code{use_ggplot==TRUE}.}

\item{...}{further arguments.}
}
\description{
Mirror of \code{\link{qua}}: probability distribution fitting with L-Moments
}
\examples{

library(RMAWGEN)


data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
                       PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

y <- annual.agg(x,dd=1:5,time) 
lmom <- annual.agg.samlmu(y)
lmrd(lmom)
z <- annual.agg.pel(distrib="gpa",x=y,lmom=lmom)

out <- annual.agg.qua(para=z)

\donttest{

library(GSODR)
library(magrittr)
library(data.table)
library(dplyr)

set.seed(123)

years <- 1937:2020
gsod <- get_GSOD(years=years,station="623180-99999") ##ALEXANDRIA INTL EG 623180-99999
prec <- gsod \%>\% select(YEARMODA,PRCP,PRCP_ATTRIBUTES) \%>\%
 mutate(YEARMODA=as.Date(YEARMODA,format="\%Y-\%m-\%d")) 

dds <- range(prec$YEARMODA)
## See GSODR documentation
yymmdds <- seq(from=dds[1],to=dds[2],by="day")
prec <- data.table::data.table(YEARMODA=yymmdds) \%>\% full_join(prec)
y <- annual.agg(x=prec$PRCP,dd=1:5,time=prec$YEARMODA)
### y$aggr cannot be -Inf or +Inf
y$aggr[y$aggr==-Inf] <- NA
lmom <- annual.agg.samlmu(y)
lmrd(lmom)
z <- annual.agg.pel(distrib="gev",x=y,lmom=lmom)
out <- annual.agg.qua(f=c(1:49)/50,para=z)
}
}
