% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2vec.R
\name{df2vec}
\alias{df2vec}
\title{Data frame (table) to vector}
\usage{
df2vec(
  x,
  aggr.name = "aggr",
  index.name = "index",
  dd.name = "dd",
  dd_formatter = "D\%03d"
)
}
\arguments{
\item{x}{object returned by \code{\link{annual.agg}} or \code{\link{yearly.agg}}}

\item{aggr.name, dd.name, index.name}{optional column mane for \code{x}. See function usage.}

\item{dd_formatter}{string formatter for duration in the function value}
}
\description{
Data frame (table) to vector
}
\examples{

## TO DO
#' 
library(RMAWGEN)
data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

out <- annual.agg(x,time)
out2 <- df2vec(out)
out3 <- vec2df(out2)

}
\seealso{
\code{\link{vec2df}}
}
