% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2df.R
\name{vec2df}
\alias{vec2df}
\title{Vector to Data frame (table) or Tibble}
\usage{
vec2df(
  x,
  nn = names(x),
  aggr.name = "aggr",
  index.name = "index",
  dd.name = "dd",
  dd_formatter = "D\%03d",
  numeric_dd = TRUE,
  numeric_index = FALSE,
  return_tibble = TRUE
)
}
\arguments{
\item{x}{object returned by \code{\link{annual.agg}} or \code{\link{yearly.agg}}}

\item{nn}{names of \code{x}}

\item{aggr.name, dd.name, index.name}{optional column mane for \code{x}. See function usage.}

\item{dd_formatter}{string formatter for duration in the function value}

\item{numeric_dd}{logical. Default is \code{TRUE}.}

\item{numeric_index}{logical. Default is \code{FALSE}.}

\item{return_tibble}{logical}
}
\description{
Vector to Data frame (table) or Tibble
}
\examples{

## TO DO
#' 
library(RMAWGEN)
data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

out <- annual.agg(x,time)
out2 <- df2vec(out)
out3 <- vec2df(out2)

}
\seealso{
\code{\link{df2vec}}
}
