% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.agg.idf.samlmu.R
\name{annual.agg.idf.samlmu}
\alias{annual.agg.idf.samlmu}
\alias{yearly.agg.idf.samlmu}
\title{Sample L-Moments of annual maxima or minima values}
\usage{
annual.agg.idf.samlmu(
  x,
  lmom = NULL,
  aggr.name = "aggr",
  dd.name = "dd",
  nn_moms = c("l_1", "l_2"),
  use_ggplot = TRUE,
  col = "blue",
  like_lmom = TRUE,
  dd_formatter = "\%03d",
  ...
)

yearly.agg.idf.samlmu(x, ...)
}
\arguments{
\item{x}{object returned by \code{\link{annual.agg}} or \code{\link{yearly.agg}} or a vector which will be processed through \code{\link{vec2df}}}

\item{lmom}{object returned by \code{\link{annual.agg.samlmu}} or \code{NULL}.}

\item{aggr.name, dd.name}{optional column mane for \code{x}. See function usage.}

\item{nn_moms}{column names of \code{lmom} whose l-moments  are scaled with duration event.}

\item{use_ggplot}{logical. if it is \code{TRUE} (default) regression plot is created. See function result descriptions.}

\item{col}{argument for \code{ggplot2} functions. It is used if \code{use_ggplot==TRUE}.}

\item{like_lmom}{logical. If it is \code{TRUE}, function returns a data frame formatted like \code{lmom} to be used for \code{\link{annual.agg.pel}}, otherwise it returns the outcome of \code{\link{samlmu}} .}

\item{dd_formatter}{argument used by \code{\link{vec2df}}}

\item{...}{further arguments for \code{\link{samlmu}}}
}
\value{
a data frame containing the L-moment of rainfall intensity for each duration category. Plus attributes, see function code.
}
\description{
Sample L-Moments of annual maxima or minima values
}
\examples{

library(lmomIDF)
library(RMAWGEN)

data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
                        PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

y <- annual.agg(x,dd=1:5,time)
out <- annual.agg.idf.samlmu(y)
y1 <- df2vec(y)
out1 <- annual.agg.idf.samlmu(y1)

y2 <- y1
y2[] <- NA
out2 <- annual.agg.idf.samlmu(y2)

if (max(abs(out1-out),na.rm=TRUE)>1e-8) stop("Something went wrong!") 
for (it in names(attributes(out))) {
   if (it=="gg") { ## TO CHECK
   } else if (is.list(attr(out,it))) for (itn in names(attr(out,it))) {
         if (!identical(attr(out,it)[[itn]],attr(out1,it)[[itn]]) & !(itn \%in\% c("model","family","terms"))) { ## TO CHECK
            msg <- sprintf("Something went wrong in \%s  \%s!",it,itn)
            stop(msg)
         }
   } else if (!identical(attr(out,it),attr(out1,it))) {
      msg <- sprintf("Something went wrong in \%s!",it)
      stop(msg)
   
   }
     
  
}





out_pel <- annual.agg.pel(distrib="gpa",x=y,lmom=out)
set.seed(560)
f=c(0.5,0.8,0.9,0.95,0.98,0.99,0.999,0.9999)
out_qua <- annual.agg.qua(para=out_pel,f=f,remove_distrib_from_boxplot=TRUE)

\donttest{

####
library(GSODR)
library(dplyr)

set.seed(123)

years <- 1937:2020
gsod <- get_GSOD(years=years,station="623180-99999") ##ALEXANDRIA INTL EG 623180-99999
prec <- gsod \%>\% select(YEARMODA,PRCP,PRCP_ATTRIBUTES) \%>\%
  mutate(YEARMODA=as.Date(YEARMODA,format="\%Y-\%m-\%d"))

dds <- range(prec$YEARMODA)
## See GSODR documentation
yymmdds <- seq(from=dds[1],to=dds[2],by="day")
prec <- data.table::data.table(YEARMODA=yymmdds) \%>\% full_join(prec)
y <- annual.agg(x=prec$PRCP,dd=1:5,time=prec$YEARMODA)
### y$aggr cannot be -Inf or +Inf
y$aggr[y$aggr==-Inf] <- NA
lmom0 <- annual.agg.samlmu(y)
lmrd(lmom0)
lmom <- annual.agg.idf.samlmu(y)


z <- annual.agg.pel(distrib="gev",x=y,lmom=lmom)
out_qua_gsod <- annual.agg.qua(f=f,para=z,remove_distrib_from_boxplot=TRUE)
attr(out_qua_gsod,"idf")
attr(out_qua_gsod,"ddf")
}



}
\seealso{
\code{\link{samlmu}},\code{\link{annual.agg.samlmu}}
}
