% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.agg.pel.R
\name{annual.agg.pel}
\alias{annual.agg.pel}
\alias{annual.agg.pel_x}
\alias{annual.agg.pel_lmom}
\alias{yearly.agg.pel}
\alias{yearly.agg.pel_x}
\alias{yearly.agg.pel_lmom}
\title{Mirror of \code{\link{pel...}}: probability distribution fitting with L-Moments}
\usage{
annual.agg.pel(
  distrib = c("exp", "gam", "gev", "glo", "gpa", "gno", "gum", "kap", "ln3", "nor",
    "pe3", "wak", "wei")[3],
  x,
  lmom = lmom,
  dd.name = "dd",
  dd_formatter = "D\%03d",
  aggr.name = "aggr",
  alternative = c("two.sided", "less", "greater"),
  exact = NULL,
  ...
)

annual.agg.pel_x(x, ...)

annual.agg.pel_lmom(lmom, ...)

yearly.agg.pel(...)

yearly.agg.pel_x(x, ...)

yearly.agg.pel_lmom(lmom, ...)
}
\arguments{
\item{distrib}{character string incating the probability distribution to fit}

\item{x}{object returned by \code{\link{annual.agg}} or \code{\link{yearly.agg}}}

\item{lmom}{object returned by \code{\link{annual.agg.samlmu}} or \code{\link{yearly.agg.samlmu}} or \code{\link{annual.agg.idf.samlmu}}}

\item{dd_formatter}{string formatter for duration in the function value}

\item{aggr.name, dd.name}{optional column names for \code{x} and/or \code{lmom}. See function usage.}

\item{alternative, exact}{arguments for \code{\link{ks.test}}}

\item{...}{further arguments}
}
\description{
Mirror of \code{\link{pel...}}: probability distribution fitting with L-Moments
}
\examples{
library(RMAWGEN)
data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
                       PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

y <- annual.agg(x,dd=1:5,time) 
lmom <- annual.agg.samlmu(y)
lmrd(lmom)

out <- annual.agg.pel(distrib="gpa",x=y,lmom=lmom)

}
