% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.agg.qua.R
\name{annual.agg.qua}
\alias{annual.agg.qua}
\alias{yearly.agg.qua}
\title{Mirror of \code{\link{qua}}: probability distribution fitting with L-Moments}
\usage{
annual.agg.qua(
  f = (0:10)/10,
  para,
  f.name = "f",
  aggr.name = NA,
  dd.name = NA,
  use_ggplot = TRUE,
  xlab = "duration",
  ylab = "value",
  fill = c("#ca0020", "#0571b0", "#bababa", "#bababb", "#bababc"),
  nrun = 5,
  ...
)

yearly.agg.qua(...)
}
\arguments{
\item{f}{vector of probabilities/frequencies.}

\item{para}{object returned by \code{\link{annual.agg.pel}}.}

\item{f.name}{column name for the probabilities/frequancies. Deafault is \code{"f"}.}

\item{aggr.name, dd.name}{column names. see \code{\link{annual.agg.samlmu}} or \code{\link{annual.agg.pel}}.}

\item{use_ggplot}{logical. if iy is \code{TRUE} (dafault) boxplots are created. See function result descriptions.}

\item{xlab, ylab, fill}{arguments for \code{ggplot2} functions. It is used if \code{use_ggplot==TRUE}. See \code{\link{geom_boxplot}},\code{\link{scale_fill_manual}},\code{\link{labs}}}

\item{nrun}{additional Monte-Carlo generations. It is used if \code{use_ggplot==TRUE}.}

\item{...}{further arguments.}
}
\description{
Mirror of \code{\link{qua}}: probability distribution fitting with L-Moments
}
\examples{
library(RMAWGEN)
set.seed(123)
##
data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
                       PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

y <- annual.agg(x,dd=1:5,time) 
lmom <- annual.agg.samlmu(y)
lmrd(lmom)
z <- annual.agg.pel(distrib="gpa",x=y,lmom=lmom)

out <- annual.agg.qua(para=z)
}
