% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.agg.samlmu.R
\name{annual.agg.samlmu}
\alias{annual.agg.samlmu}
\alias{yearly.agg.samlmu}
\title{Sample L-Moments of annual maxima or minima values}
\usage{
annual.agg.samlmu(x, aggr.name = "aggr", dd.name = "dd", ...)

yearly.agg.samlmu(x, ...)
}
\arguments{
\item{x}{object returned by \code{\link{annual.agg}} or \code{\link{yearly.agg}}}

\item{aggr.name, dd.name}{optional column mane for \code{x}. See function usage.}

\item{...}{further arguments for \code{\link{samlmu}}}
}
\description{
Sample L-Moments of annual maxima or minima values
}
\examples{

library(RMAWGEN)
data(trentino)
time <- as.Date(sprintf("\%04d-\%02d-\%02d",PRECIPITATION$year,
PRECIPITATION$month,PRECIPITATION$day),format="\%Y-\%m-\%d")
x <- PRECIPITATION$B8570

y <- annual.agg(x,dd=1:10,time)
out <- annual.agg.samlmu(y)

lmrd(out)

summary(lm(log(l_1) ~ log(dd),data=out))

}
